tests/wpt/web-platform-tests/fenced-frame/navigator-keyboard-layout-map.https.html

<!DOCTYPE html>
<title>Test navigator.keyboard.getLayoutMap()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<script src="/common/utils.js"></script>

<body>

<script>
promise_test(async () => {
  const keyboard_layout_key = token();

  attachFencedFrame(generateURL(
      "resources/navigator-keyboard-layout-map-inner.html",
      [keyboard_layout_key]));
  const actual_result = await nextValueFromServer(keyboard_layout_key);

  assert_equals(actual_result, "rejected",
                "The fenced frame is not allowed to fetch keyboard map.");

}, "keyboard.getLayoutMap()");
</script>

</body>
